Add params based namespaces in ImplicitBinder
This is done to allow for even the following method calls:

- implicitBinder.ScanForAnnotatedClasses ("Some.Namespace");
- implicitBinder.ScanForAnnotatedClasses (“Some.A”, “Other.B”);

In the previous implementation, one had to write:

implicitBinder.ScanForAnnotatedClasses (new string[] { “Some.A” });

Now both types of formats are supported.

Tests haven’t been added, for the behaviour that params allow is
guaranteed by the language specification itself.

When no namespaces are provided then the behaviour is same as passing
an empty string array.

This closes #170.